# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.CloudUserStatus:
    enum:
    - active
    - inactive
    - deleted
    type: string
    x-enum-varnames:
    - CloudUserStatusActive
    - CloudUserStatusInactive
    - CloudUserStatusDeleted
  domain.CloudUserType:
    enum:
    - api_key
    - access_key
    - ram_user
    - iam_user
    - cam_user
    - volc_user
    type: string
    x-enum-comments:
      CloudUserTypeCAMUser: 腾讯云 CAM 用户
      CloudUserTypeIAMUser: AWS/华为云 IAM 用户
      CloudUserTypeRAMUser: 阿里云 RAM 用户
      CloudUserTypeVolcUser: 火山云用户
    x-enum-descriptions:
    - ""
    - ""
    - 阿里云 RAM 用户
    - AWS/华为云 IAM 用户
    - 腾讯云 CAM 用户
    - 火山云用户
    x-enum-varnames:
    - CloudUserTypeAPIKey
    - CloudUserTypeAccessKey
    - CloudUserTypeRAMUser
    - CloudUserTypeIAMUser
    - CloudUserTypeCAMUser
    - CloudUserTypeVolcUser
...
  domain.SyncTargetType:
    enum:
    - user
    - group
    - account
    type: string
    x-enum-comments:
      SyncTargetTypeAccount: 云账号级别同步
    x-enum-descriptions:
    - ""
    - ""
    - 云账号级别同步
    x-enum-varnames:
    - SyncTargetTypeUser
    - SyncTargetTypeGroup
    - SyncTargetTypeAccount
  domain.SyncTaskType:
    enum:
    - user_sync
    - permission_sync
    - group_sync
    - batch_user_sync
    type: string
    x-enum-comments:
      SyncTaskTypeBatchUserSync: 批量用户同步（从云平台拉取所有用户）
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - 批量用户同步（从云平台拉取所有用户）
    x-enum-varnames:
    - SyncTaskTypeUserSync
    - SyncTaskTypePermissionSync
    - SyncTaskTypeGroupSync
    - SyncTaskTypeBatchUserSync
...
  ginx.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  github_com_Havens-blog_e-cam-service_internal_shared_domain.CloudProvider:
    enum:
    - aliyun
    - aws
    - azure
    - tencent
    - huawei
    - volcano
    - volcengine
    type: string
    x-enum-comments:
      CloudProviderAWS: Amazon Web Services
      CloudProviderAliyun: 阿里云
      CloudProviderAzure: Microsoft Azure
      CloudProviderHuawei: 华为云
      CloudProviderTencent: 腾讯云
      CloudProviderVolcano: 火山云 (别名)
      CloudProviderVolcengine: 火山引擎
    x-enum-descriptions:
    - 阿里云
    - Amazon Web Services
    - Microsoft Azure
    - 腾讯云
    - 华为云
    - 火山云 (别名)
    - 火山引擎
    x-enum-varnames:
    - CloudProviderAliyun
    - CloudProviderAWS
    - CloudProviderAzure
    - CloudProviderTencent
    - CloudProviderHuawei
    - CloudProviderVolcano
    - CloudProviderVolcengine
...
  github_com_Havens-blog_e-cam-service_internal_shared_domain.SyncTask:
    properties:
      cloud_account_id:
        type: integer
      create_time:
        type: string
      ctime:
        type: integer
      end_time:
        type: string
      error_message:
        type: string
      id:
        type: integer
      max_retries:
        type: integer
      progress:
        type: integer
      provider:
        $ref: '#/definitions/github_com_Havens-blog_e-cam-service_internal_shared_domain.CloudProvider'
      retry_count:
        type: integer
      start_time:
        type: string
      status:
        $ref: '#/definitions/github_com_Havens-blog_e-cam-service_internal_shared_domain.SyncTaskStatus'
      target_id:
        type: integer
      target_type:
        $ref: '#/definitions/domain.SyncTargetType'
      task_type:
        $ref: '#/definitions/domain.SyncTaskType'
      update_time:
        type: string
      utime:
        type: integer
    type: object
  github_com_Havens-blog_e-cam-service_internal_shared_domain.SyncTaskStatus:
    enum:
    - pending
    - running
    - success
    - failed
    - retrying
    type: string
    x-enum-varnames:
    - SyncTaskStatusPending
    - SyncTaskStatusRunning
    - SyncTaskStatusSuccess
    - SyncTaskStatusFailed
    - SyncTaskStatusRetrying
...
  web.AssignUserGroupsVO:
    properties:
      group_ids:
        items:
          type: integer
        minItems: 1
        type: array
      user_ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - group_ids
    - user_ids
    type: object
...
  web.CreateEnvReq:
    properties:
      code:
        type: string
      color:
        type: string
      description:
        type: string
      name:
        type: string
      order:
        type: integer
    required:
    - code
    - name
    type: object
...
  web.CreateUserVO:
    properties:
      cloud_account_id:
        type: integer
      display_name:
        maxLength: 200
        type: string
      email:
        type: string
      tenant_id:
        type: string
      user_groups:
        description: 用户组ID列表
        items:
          type: integer
        type: array
      user_type:
        $ref: '#/definitions/domain.CloudUserType'
      username:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - cloud_account_id
    - tenant_id
    - user_type
    - username
    type: object
...
  web.EnvironmentVO:
    properties:
      code:
        type: string
      color:
        type: string
      create_time:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      order:
        type: integer
      status:
        type: integer
      update_time:
        type: integer
    type: object
...
  web.PageResult:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  web.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  web.UpdateEnvReq:
    properties:
      code:
        type: string
      color:
        type: string
      description:
        type: string
      name:
        type: string
      order:
        type: integer
      status:
        type: integer
    required:
    - code
    - name
    type: object
...
  web.UpdateUserVO:
    properties:
      display_name:
        type: string
      email:
        type: string
      status:
        $ref: '#/definitions/domain.CloudUserStatus'
      user_groups:
        description: 用户组ID列表
        items:
          type: integer
        type: array
    type: object
host: localhost:8001
info:
  contact:
    email: support@example.com
    name: API Support
  description: 企业云资产管理服务 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: E-CAM Service API
  version: "1.0"
paths:
...
  /api/v1/cam/iam/permissions/users/{user_id}:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 获取用户的所有权限
      tags:
      - 权限管理
...
  /api/v1/cam/iam/users:
    get:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 云厂商
        in: query
        name: provider
        type: string
      - description: 用户类型
        in: query
        name: user_type
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.PageResult'
      summary: 查询用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建用户请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.CreateUserVO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 创建云用户
      tags:
      - 用户管理
  /api/v1/cam/iam/users/{id}:
    delete:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 删除用户
      tags:
      - 用户管理
    get:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 获取用户详情
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新用户请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.UpdateUserVO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 更新用户信息
      tags:
      - 用户管理
  /api/v1/cam/iam/users/batch-assign:
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 批量分配用户组请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.AssignUserGroupsVO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 批量分配用户组
      tags:
      - 用户管理
  /api/v1/cam/iam/users/sync:
    post:
      description: 创建异步同步任务，立即返回任务ID，可通过任务ID查询同步进度
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 云账号ID
        in: query
        name: cloud_account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Result'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Havens-blog_e-cam-service_internal_shared_domain.SyncTask'
              type: object
      summary: 同步云平台用户（异步）
      tags:
      - 用户管理
...
  /api/v1/cam/service-tree/environments:
    get:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.EnvironmentVO'
                  type: array
              type: object
      summary: 获取环境列表
      tags:
      - 服务树-环境
    post:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 环境信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.CreateEnvReq'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  format: int64
                  type: integer
              type: object
      summary: 创建环境
      tags:
      - 服务树-环境
  /api/v1/cam/service-tree/environments/{id}:
    delete:
      parameters:
      - description: 环境ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 删除环境
      tags:
      - 服务树-环境
    get:
      parameters:
      - description: 环境ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  $ref: '#/definitions/web.EnvironmentVO'
              type: object
      summary: 获取环境详情
      tags:
      - 服务树-环境
    put:
      parameters:
      - description: 环境ID
        in: path
        name: id
        required: true
        type: integer
      - description: 环境信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.UpdateEnvReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 更新环境
      tags:
      - 服务树-环境
  /api/v1/cam/service-tree/environments/init:
    post:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 初始化默认环境
      tags:
      - 服务树-环境
...
schemes:
- http
- https
swagger: "2.0"